First stab at writing functions to put obs op in BSR form.
The influence functions are really big; with the savings in the
background error covariance elsewhere in this package, they are the
largest.  For regional inversions, we only need a few weeks of
transport before everything leaves the domain.  This tries to keep the
space savings in the inversion, not just on disk.