Include a description for 'Helpers for real-data inversions'.
Just including that title doesn't actually help anyone who doesn't
already know what's going on.  This isn't a complete description of
what they're for and why someone might want to use them, but it's a
start.

The project really needs a better introduction to how to use the code
than "just look at the three thousand lines of tests, each of which
does a piece of what you want to do, but without explanation, and the
thousand lines of extremely messy code I've been tinkering with for a
year and a half through a couple of big changes in how the code works,
which occasionally has some explanation that might still be valid."
API documentation only really works once you know what's going on.

That or make a main script that reads a config file and performs the
inversion described.  The wrapper code is a first step in that
direction, but I don't like the interface terribly much.