WIP: Try to find determinant of HomogeneousIsotropicCorrelations.
The primary data stored by this class is a list of eigenvalues, so
this should be easy.  For some reason it's broken for all but trivial
cases.  I can get the determinant to be one for diagonal correlations,
but this is not the primary class for diagonal operators, especially
not ones proportional to the identity matrix.  I have no idea why this
doesn't work.  This may be enough for the purposes of this package,
since it's usually only off by 50% or so, but I really don't like
calling the result a `determinant`.